requirements.txt
psutil==3.3.0
kombu==3.0.33
pymongo==3.2
amqp==1.4.9
pyyaml==3.11
redis==2.10.5
tornado==4.2.1
flask==0.11
gevent==1.1.1